backend/requirements.txt
Flask==3.1.0
Flask-Cors==5.0.1
Flask-JWT-Extended==4.7.1
Flask-Migrate==4.1.0
Flask-SQLAlchemy==3.1.1
SQLAlchemy==2.0.39
alembic==1.15.1
celery==5.4.0
psycopg[binary]==3.2.6
python-dotenv==1.0.1
requests==2.32.5
redis==5.2.1
stripe==12.0.0
openai==1.68.2
yfinance==1.2.0
pytest==8.3.5
